change state name & move component to root
